[api] rename pole-emplois -> pole-emploi
