+++ api/lib/application/pole-emploi/index.js
@@ -7,7 +7,7 @@ exports.register = async function (server) {
   server.route([
     {
       method: 'POST',
-      path: '/api/pole-emplois/users',
+      path: '/api/pole-emploi/users',
       config: {
         auth: false,
         handler: poleEmploiController.createUser,
@@ -64,4 +64,4 @@ exports.register = async function (server) {
   ]);
 };
 
-exports.name = 'pole-emplois-api';
+exports.name = 'pole-emploi-api';
