+++ api/lib/routes.js
@@ -30,7 +30,7 @@ module.exports = [
   require('./application/organization-invitations'),
   require('./application/organizations'),
   require('./application/passwords'),
-  require('./application/pole-emplois'),
+  require('./application/pole-emploi'),
   require('./application/prescribers'),
   require('./application/progressions'),
   require('./application/saml'),
