+++ api/tests/unit/application/pole-emploi/index_test.js
@@ -4,8 +4,8 @@ const {
   sinon,
   generateValidRequestAuthorizationHeaderForApplication,
 } = require('../../../test-helper');
-const poleEmploiController = require('../../../../lib/application/pole-emplois/pole-emploi-controller');
-const moduleUnderTest = require('../../../../lib/application/pole-emplois');
+const poleEmploiController = require('../../../../lib/application/pole-emploi/pole-emploi-controller');
+const moduleUnderTest = require('../../../../lib/application/pole-emploi');
 
 describe('Unit | Router | pole-emploi-router', function () {
   describe('GET /api/pole-emploi/envois', function () {
